requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
Pillow==10.0.1
Werkzeug==2.3.7
